requirements.txt
# Основные библиотеки для работы с Telegram
telebot

# Библиотеки для работы с базами данных
pymongo                   # Для работы с MongoDB
motor                     # Асинхронный драйвер для MongoDB (если нужен)
#SQLAlchemy                # Если планируете использовать SQL базы данных

# Библиотеки для работы с HTTP запросами
aiohttp                   # Асинхронные HTTP запросы

# Библиотеки для работы с Excel
xlwt                      # Для записи в Excel файлы (если необходимо)

# Библиотека для скачивания видео с YouTube
#yt-dlp                    # Для скачивания видео с YouTube

# Дополнительные библиотеки
python-dotenv             # Для работы с .env файлами
#loguru                    # Удобный логгер (если нужен)
